removed: .fleet folder

+++ act.sh
@@ -1,14 +1,19 @@
 #!/bin/bash
 
-if [ ! -f bin/act ]; then
-  cd .fleet
-  curl -sL https://raw.githubusercontent.com/nektos/act/master/install.sh -o install.sh
-  chmod +x install.sh
-  echo "Acquiring sudo to be able to install act..."
-  sudo ./install.sh
-  rm install.sh
-  cd ..
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
 fi
 
-echo "Acquiring sudo to be able to run docker..."
-sudo .fleet/bin/act
+cd "$(dirname "$0")"
+
+if [ ! -f bin/act ]; then
+  cd .github/scripts
+  curl -sL https://raw.githubusercontent.com/nektos/act/master/install.sh -o install.sh
+  chmod +x install.sh
+  ./install.sh
+  rm install.sh
+  cd ../..
+fi
+
+.github/scripts/bin/act
